refactor: db 모듈 생성
