+++ db/__init__.py
@@ -1,0 +1,14 @@
+"""
+Database modules for MangMangDae-AI project.
+
+This package contains database connection and operation classes:
+- DynamoDB: AWS DynamoDB operations
+- Pinecone: Pinecone vector database operations  
+- Postgres: PostgreSQL database operations
+"""
+
+from .dynamodb import DynamoDB
+from .pinecone_db import Pinecone
+from .postgres import Postgres
+
+__all__ = ['DynamoDB', 'Pinecone', 'Postgres']
